Remove HTTP/2 workaround for pygsite
